Fixed some issues discussed in PR #48: - Invalid frame correctly referenced - Fixed a missing import in setup of Global Time Keeper - Removed the custom BVTKException and moved bvtk_assert to the core - animation.py moved to the main directory and renamed animation_helper.py - Fixed a bug where the Global Time Keeper could overwrite changes in Time Selector nodes with active 'Use Scene Time'
