+++ customfilter.py
@@ -1,7 +1,7 @@
 from .core import l # Import logging
 from .core import *
-from .time.animation import AnimationHelper
-from .cache import PersistentStorageUser
+from .animation_helper import AnimationHelper
+from .cache import PersistentStorageUser, persistent_storage
 
 # -----------------------------------------------------------------------------
 # Custom filter
